Lab2_Colliders
Notes:

- Created an example scene complete with a background, and some tiles
- Added 2D Box Colliders to the scene
- Note: The platforms are not intended to be jumped through

- Added a player character (Samus)
- Added 2D Box Collider and 2D Rigidbody
- Added default idle animation
Note: There were far too many sprites included in the sprite sheet for each of the Samus suits included. I went through and named the ones that I intend to use, as well as a few that I will use if time is permitting. There are some sprite frames that have the default name applied to them, and this was simply because there were too many to name, and I couldn't figure out what they were all for.